Display task list as table
Rather than generating a ul element with li elements for each task, the task list now generates a table with columns for the task title, task due date, and a completion button. The button--as does the completion button for overall projects--does not seem to accurately assign 'completed: true/false' and the CSS styles appropriate; fixing this is the next priority.