Fix rendering of completed projects as incomplete
Upon each new rendering of the list of projects, it removed the 'completed=true' attribute from any complete projects, causing them to be displayed as incomplete even though examination of the project object revealed they were still 'completed' under the hood. There is now a check at the beginning of displayProjectList() that adds the attribute if the project has 'project.completed == true'.