Move task table generation to separate function; fix incorrect display of complete tasks
Moved the generation of the table of tasks for each project to its own function. Also fixed the display of completed tasks; they were having the same issue as completed projects, where the addition of a new task/project caused the re-rendering of the whole list of tasks/projects without the 'completed=true' attribute. Fixed the issue for tasks in the same way by adding a check for completion before the rendering.